requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
discord-py-interactions==4.3.4
discord.py==2.0.1
exceptiongroup==1.0.1
frozenlist==1.3.1
h11==0.14.0
idna==3.4
interactions-files==1.1.5
multidict==6.0.2
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
yarl==1.8.1
